requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
djangorestframework-simplejwt>=5.4.0
Pillow>=10.0.0
python-dateutil>=2.9.0
